Include optimal bet

+++ src/main/java/io/freezing/ai/bot/action/impl/RaiseAction.java
@@ -4,10 +4,10 @@ import io.freezing.ai.bot.BotActionRationale;
 import io.freezing.ai.bot.action.BotAction;
 
 public class RaiseAction implements BotAction {
-    private final int amount;
+    private final double amount;
     private final BotActionRationale rationale;
 
-    public RaiseAction(int amount, BotActionRationale rationale) {
+    public RaiseAction(double amount, BotActionRationale rationale) {
         this.amount = amount;
         this.rationale = rationale;
     }
@@ -17,12 +17,12 @@ public class RaiseAction implements BotAction {
         return rationale;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
     @Override
     public String toString() {
-        return String.format("RaiseAction(amount = %d, rationale = %s)", amount, rationale);
+        return String.format("RaiseAction(amount = %f, rationale = %s)", amount, rationale);
     }
 }
